reverse the dependency direction between State and EventId
