+++ module/parser-base/build.gradle.kts
@@ -12,8 +12,6 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                api(project(":module:renderer-state"))
-
                 implementation(libs.androidx.annotation)
             }
         }
